refactor: move old stuff into /src/old
Makes it clear which parts are the new lib stuff and which are the old routes.
