+++ src/old/api/columns.ts
@@ -1,11 +1,11 @@
 import { Router } from 'express'
 import format, { ident, literal } from 'pg-format'
 import SQL from 'sql-template-strings'
-import { RunQuery } from '../lib/connectionPool'
-import sql = require('../lib/sql')
-import { DEFAULT_SYSTEM_SCHEMAS } from '../lib/constants'
-import { Tables } from '../lib/interfaces'
-import logger from '../server/logger'
+import { RunQuery } from '../connectionPool'
+import sql = require('../../lib/sql')
+import { DEFAULT_SYSTEM_SCHEMAS } from '../../lib/constants'
+import { Tables } from '../interfaces'
+import logger from '../../server/logger'
 
 /**
  * @param {boolean} [include_system_schemas=false] - Return system schemas as well as user schemas
