+++ src/old/api/config.ts
@@ -1,8 +1,8 @@
 import { Router } from 'express'
 
-import { RunQuery } from '../lib/connectionPool'
-import sql = require('../lib/sql')
-import logger from '../server/logger'
+import { RunQuery } from '../connectionPool'
+import sql = require('../../lib/sql')
+import logger from '../../server/logger'
 const { configSql, versionSql } = sql
 
 const router = Router()
