+++ src/old/api/policies.ts
@@ -1,11 +1,11 @@
 import { Router } from 'express'
 import format from 'pg-format'
-import { toTransaction } from '../lib/helpers'
-import { RunQuery } from '../lib/connectionPool'
-import { DEFAULT_SYSTEM_SCHEMAS } from '../lib/constants'
-import { Tables } from '../lib/interfaces'
-import sqlTemplates = require('../lib/sql')
-import logger from '../server/logger'
+import { toTransaction } from '../../lib/helpers'
+import { RunQuery } from '../connectionPool'
+import { DEFAULT_SYSTEM_SCHEMAS } from '../../lib/constants'
+import { Tables } from '../interfaces'
+import sqlTemplates = require('../../lib/sql')
+import logger from '../../server/logger'
 
 /**
  * @param {string} [include_system_schemas=false] - Return system schemas as well as user schemas
