+++ src/old/api/publications.ts
@@ -1,8 +1,8 @@
 import { Router } from 'express'
 import { ident, literal } from 'pg-format'
-import sql = require('../lib/sql')
-import { RunQuery } from '../lib/connectionPool'
-import logger from '../server/logger'
+import sql = require('../../lib/sql')
+import { RunQuery } from '../connectionPool'
+import logger from '../../server/logger'
 
 const { publicationsSql } = sql
 const router = Router()
