+++ src/old/api/types.ts
@@ -1,11 +1,11 @@
 import { Router } from 'express'
 
-import sql = require('../lib/sql')
-const { functionsSql } = sql
-import { RunQuery } from '../lib/connectionPool'
-import { DEFAULT_SYSTEM_SCHEMAS } from '../lib/constants'
-import { Functions } from '../lib/interfaces'
-import logger from '../server/logger'
+import sql = require('../../lib/sql')
+const { typesSql } = sql
+import { RunQuery } from '../connectionPool'
+import { DEFAULT_SYSTEM_SCHEMAS } from '../../lib/constants'
+import { Types } from '../interfaces'
+import logger from '../../server/logger'
 
 /**
  * @param {boolean} [include_system_schemas=false] - Return system schemas as well as user schemas
@@ -15,12 +15,13 @@ interface QueryParams {
 }
 
 const router = Router()
+
 router.get('/', async (req, res) => {
   try {
-    const { data } = await RunQuery(req.headers.pg, functionsSql)
+    const { data } = await RunQuery(req.headers.pg, typesSql)
     const query: QueryParams = req.query
     const include_system_schemas = query?.include_system_schemas === 'true'
-    let payload: Functions.Function[] = data
+    let payload: Types.Type[] = data
     if (!include_system_schemas) payload = removeSystemSchemas(data)
     return res.status(200).json(payload)
   } catch (error) {
@@ -29,7 +30,7 @@ router.get('/', async (req, res) => {
   }
 })
 
-const removeSystemSchemas = (data: Functions.Function[]) => {
+const removeSystemSchemas = (data: Types.Type[]) => {
   return data.filter((x) => !DEFAULT_SYSTEM_SCHEMAS.includes(x.schema))
 }
